requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.7.4
python-multipart==0.0.9
Pillow==10.3.0
torch==2.2.2
torchvision==0.17.2
numpy==1.26.4
datasets==2.21.0
huggingface_hub==0.25.2
duckduckgo_search==6.1.6
requests==2.32.3
